Abstracting url and http request
