+++ cmd/checks.go
@@ -3,8 +3,15 @@ package cmd
 import (
 	"context"
 	"fmt"
+	"io/ioutil"
+	"net"
+	"net/http"
+	"net/url"
 	"os"
+	"strconv"
 
+	"github.com/dcos/dcos-checks/client"
+	"github.com/pkg/errors"
 	"github.com/sirupsen/logrus"
 )
 
@@ -43,6 +50,13 @@ type DCOSChecker interface {
 	Run(context.Context, *CLIConfigFlags) (string, int, error)
 }
 
+// URLFields is used to construct the url
+type URLFields struct {
+	host string
+	port int
+	path string
+}
+
 // RunCheck is a helper function to run the check and emit the result.
 func RunCheck(ctx context.Context, check DCOSChecker) {
 	output, retCode, err := check.Run(ctx, DCOSConfig)
@@ -55,4 +69,56 @@ func RunCheck(ctx context.Context, check DCOSChecker) {
 	}
 
 	os.Exit(retCode)
+}
+
+// HTTPRequest verifies the results of the request
+func HTTPRequest(cfg *CLIConfigFlags, urlOptions URLFields) ([]byte, error) {
+	httpClient, err := client.NewClient(cfg.IAMConfig, cfg.CACert)
+	if err != nil {
+		return nil, errors.Wrap(err, "unable to create HTTP client")
+	}
+
+	url, err := getURL(httpClient, cfg, urlOptions)
+	if err != nil {
+		return nil, err
+	}
+
+	logrus.Debugf("GET %s", url)
+	req, err := http.NewRequest("GET", url.String(), nil)
+	if err != nil {
+		return nil, errors.Wrap(err, "unable to create a new HTTP request")
+	}
+
+	resp, err := httpClient.Do(req)
+	if err != nil {
+		return nil, errors.Wrapf(err, "unable to execute GET %s", url)
+	}
+
+	responseData, err := ioutil.ReadAll(resp.Body)
+	if err != nil {
+		return nil, errors.Wrapf(err, "unable to read response body")
+	}
+	defer resp.Body.Close()
+
+	return responseData, nil
+}
+
+func getURL(httpClient *http.Client, cfg *CLIConfigFlags, urlOptions URLFields) (*url.URL, error) {
+	scheme := httpScheme
+	if cfg.ForceTLS {
+		scheme = httpsScheme
+	}
+
+	if urlOptions.port == 0 {
+		return &url.URL{
+			Scheme: scheme,
+			Host:   urlOptions.host,
+			Path:   urlOptions.path,
+		}, nil
+	}
+	return &url.URL{
+		Scheme: scheme,
+		Host:   net.JoinHostPort(urlOptions.host, strconv.Itoa(urlOptions.port)),
+		Path:   urlOptions.path,
+	}, nil
 }
